// apps/providers-service/src/__generated__/resolvers-types.d.ts
export type Scalars = {
    ID: {
        input: string;
        output: string;
    };
    String: {
        input: string;
        output: string;
    };
    Boolean: {
        input: boolean;
        output: boolean;
    };
    Int: {
        input: number;
        output: number;
    };
    Float: {
        input: number;
        output: number;
    };
    DateTime: {
        input: any;
        output: any;
    };
    _FieldSet: {
        input: any;
        output: any;
    };
};
export type Address = {
    __typename?: 'Address';
    city: Scalars['String']['output'];
    country: Scalars['String']['output'];
    state: Scalars['String']['output'];
    street: Scalars['String']['output'];
    zipCode: Scalars['String']['output'];
};
export type AddressInput = {
    city: Scalars['String']['input'];
    country: Scalars['String']['input'];
    state: Scalars['String']['input'];
    street: Scalars['String']['input'];
    zipCode: Scalars['String']['input'];
};
export type Case = {
    __typename?: 'Case';
    id: Scalars['ID']['output'];
};
export type CreateFacilityInput = {
    address: AddressInput;
    name: Scalars['String']['input'];
    phone: Scalars['String']['input'];
};
export type CreateProviderInput = {
    credentials: Scalars['String']['input'];
    email: Scalars['String']['input'];
    facilityId?: InputMaybe<Scalars['ID']['input']>;
    firstName: Scalars['String']['input'];
    lastName: Scalars['String']['input'];
    npi: Scalars['String']['input'];
    phone: Scalars['String']['input'];
    specialty: Scalars['String']['input'];
};
export type CreateVisitInput = {
    caseIds: Array<Scalars['ID']['input']>;
    chiefComplaint?: InputMaybe<Scalars['String']['input']>;
    hospitalId: Scalars['ID']['input'];
    patientId: Scalars['ID']['input'];
    providerId: Scalars['ID']['input'];
    scheduledAt: Scalars['DateTime']['input'];
    type: VisitType;
};
export type Facility = {
    __typename?: 'Facility';
    address: Address;
    id: Scalars['ID']['output'];
    name: Scalars['String']['output'];
    phone: Scalars['String']['output'];
};
export type Hospital = {
    __typename?: 'Hospital';
    id: Scalars['ID']['output'];
};
export type Mutation = {
    __typename?: 'Mutation';
    addCaseToVisit?: Maybe<Visit>;
    cancelVisit?: Maybe<Visit>;
    checkInVisit?: Maybe<Visit>;
    completeVisit?: Maybe<Visit>;
    createFacility?: Maybe<Facility>;
    createProvider?: Maybe<Provider>;
    createVisit?: Maybe<Visit>;
    removeCaseFromVisit?: Maybe<Visit>;
    startVisit?: Maybe<Visit>;
    updateProvider?: Maybe<Provider>;
    updateVisit?: Maybe<Visit>;
};
// ...
export type MutationCreateFacilityArgs = {
    input: CreateFacilityInput;
};
export type MutationCreateProviderArgs = {
    input: CreateProviderInput;
};
export type MutationCreateVisitArgs = {
    input: CreateVisitInput;
};
export type MutationRemoveCaseFromVisitArgs = {
    caseId: Scalars['ID']['input'];
    visitId: Scalars['ID']['input'];
};
export type MutationStartVisitArgs = {
    id: Scalars['ID']['input'];
};
export type MutationUpdateProviderArgs = {
    id: Scalars['ID']['input'];
    input: UpdateProviderInput;
};
export type MutationUpdateVisitArgs = {
    id: Scalars['ID']['input'];
    input: UpdateVisitInput;
};
export type Patient = {
    __typename?: 'Patient';
    id: Scalars['ID']['output'];
};
export type Provider = {
    __typename?: 'Provider';
    credentials: Scalars['String']['output'];
    email: Scalars['String']['output'];
    facility?: Maybe<Facility>;
    firstName: Scalars['String']['output'];
    id: Scalars['ID']['output'];
    lastName: Scalars['String']['output'];
    npi: Scalars['String']['output'];
    phone: Scalars['String']['output'];
    specialty: Scalars['String']['output'];
    visits: Array<Visit>;
};
export type Query = {
    __typename?: 'Query';
    facility?: Maybe<Facility>;
    provider?: Maybe<Provider>;
    providerByNpi?: Maybe<Provider>;
    providers: Array<Provider>;
    visit?: Maybe<Visit>;
    visitsByDateRange: Array<Visit>;
    visitsForCase: Array<Visit>;
    visitsForHospital: Array<Visit>;
    visitsForPatient: Array<Visit>;
    visitsForProvider: Array<Visit>;
};
// ...
export type Recommendation = {
    __typename?: 'Recommendation';
    id: Scalars['ID']['output'];
};
// ...
export type Visit = {
    __typename?: 'Visit';
    caseIds: Array<Scalars['ID']['output']>;
    chiefComplaint?: Maybe<Scalars['String']['output']>;
    completedAt?: Maybe<Scalars['DateTime']['output']>;
    duration?: Maybe<Scalars['Int']['output']>;
    hospitalId: Scalars['ID']['output'];
    id: Scalars['ID']['output'];
    notes?: Maybe<Scalars['String']['output']>;
    patientId: Scalars['ID']['output'];
    providerId: Scalars['ID']['output'];
    scheduledAt: Scalars['DateTime']['output'];
    startedAt?: Maybe<Scalars['DateTime']['output']>;
    status: VisitStatus;
    type: VisitType;
};
// ...
export type WithIndex<TObject> = TObject & Record<string, any>;
export type ResolversObject<TObject> = WithIndex<TObject>;
// ...
export type IsTypeOfResolverFn<T = {}, TContext = {}> = (obj: T, context: TContext, info: GraphQLResolveInfo) => boolean | Promise<boolean>;
// ...
export type AddressResolvers<ContextType = DataSourceContext, ParentType extends ResolversParentTypes['Address'] = ResolversParentTypes['Address']> = ResolversObject<{
    city?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
    country?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
    state?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
    street?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
    zipCode?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
    __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
}>;
export type CaseResolvers<ContextType = DataSourceContext, ParentType extends ResolversParentTypes['Case'] = ResolversParentTypes['Case']> = ResolversObject<{
    __resolveReference?: ReferenceResolver<Maybe<ResolversTypes['Case']>, ParentType, ContextType>;
    id?: Resolver<ResolversTypes['ID'], ParentType, ContextType>;
    __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
}>;
export interface DateTimeScalarConfig extends GraphQLScalarTypeConfig<ResolversTypes['DateTime'], any> {
    name: 'DateTime';
}
export type FacilityResolvers<ContextType = DataSourceContext, ParentType extends ResolversParentTypes['Facility'] = ResolversParentTypes['Facility']> = ResolversObject<{
    address?: Resolver<ResolversTypes['Address'], ParentType, ContextType>;
    id?: Resolver<ResolversTypes['ID'], ParentType, ContextType>;
    name?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
    phone?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
    __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
}>;
export type HospitalResolvers<ContextType = DataSourceContext, ParentType extends ResolversParentTypes['Hospital'] = ResolversParentTypes['Hospital']> = ResolversObject<{
    __resolveReference?: ReferenceResolver<Maybe<ResolversTypes['Hospital']>, ParentType, ContextType>;
    id?: Resolver<ResolversTypes['ID'], ParentType, ContextType>;
    __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
}>;
export type MutationResolvers<ContextType = DataSourceContext, ParentType extends ResolversParentTypes['Mutation'] = ResolversParentTypes['Mutation']> = ResolversObject<{
    addCaseToVisit?: Resolver<Maybe<ResolversTypes['Visit']>, ParentType, ContextType, RequireFields<MutationAddCaseToVisitArgs, 'caseId' | 'visitId'>>;
    cancelVisit?: Resolver<Maybe<ResolversTypes['Visit']>, ParentType, ContextType, RequireFields<MutationCancelVisitArgs, 'id'>>;
    checkInVisit?: Resolver<Maybe<ResolversTypes['Visit']>, ParentType, ContextType, RequireFields<MutationCheckInVisitArgs, 'id'>>;
    completeVisit?: Resolver<Maybe<ResolversTypes['Visit']>, ParentType, ContextType, RequireFields<MutationCompleteVisitArgs, 'id'>>;
    createFacility?: Resolver<Maybe<ResolversTypes['Facility']>, ParentType, ContextType, RequireFields<MutationCreateFacilityArgs, 'input'>>;
    createProvider?: Resolver<Maybe<ResolversTypes['Provider']>, ParentType, ContextType, RequireFields<MutationCreateProviderArgs, 'input'>>;
    createVisit?: Resolver<Maybe<ResolversTypes['Visit']>, ParentType, ContextType, RequireFields<MutationCreateVisitArgs, 'input'>>;
    removeCaseFromVisit?: Resolver<Maybe<ResolversTypes['Visit']>, ParentType, ContextType, RequireFields<MutationRemoveCaseFromVisitArgs, 'caseId' | 'visitId'>>;
    startVisit?: Resolver<Maybe<ResolversTypes['Visit']>, ParentType, ContextType, RequireFields<MutationStartVisitArgs, 'id'>>;
    updateProvider?: Resolver<Maybe<ResolversTypes['Provider']>, ParentType, ContextType, RequireFields<MutationUpdateProviderArgs, 'id' | 'input'>>;
    updateVisit?: Resolver<Maybe<ResolversTypes['Visit']>, ParentType, ContextType, RequireFields<MutationUpdateVisitArgs, 'id' | 'input'>>;
}>;
export type PatientResolvers<ContextType = DataSourceContext, ParentType extends ResolversParentTypes['Patient'] = ResolversParentTypes['Patient']> = ResolversObject<{
    __resolveReference?: ReferenceResolver<Maybe<ResolversTypes['Patient']>, ParentType, ContextType>;
    id?: Resolver<ResolversTypes['ID'], ParentType, ContextType>;
    __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
}>;
export type ProviderResolvers<ContextType = DataSourceContext, ParentType extends ResolversParentTypes['Provider'] = ResolversParentTypes['Provider']> = ResolversObject<{
    __resolveReference?: ReferenceResolver<Maybe<ResolversTypes['Provider']>, {
        __typename: 'Provider';
    } & GraphQLRecursivePick<ParentType, {
        "id": true;
    }>, ContextType>;
    credentials?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
    email?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
    facility?: Resolver<Maybe<ResolversTypes['Facility']>, ParentType, ContextType>;
    firstName?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
    id?: Resolver<ResolversTypes['ID'], ParentType, ContextType>;
    lastName?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
    npi?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
    phone?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
    specialty?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
    visits?: Resolver<Array<ResolversTypes['Visit']>, ParentType, ContextType>;
    __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
}>;
export type QueryResolvers<ContextType = DataSourceContext, ParentType extends ResolversParentTypes['Query'] = ResolversParentTypes['Query']> = ResolversObject<{
    facility?: Resolver<Maybe<ResolversTypes['Facility']>, ParentType, ContextType, RequireFields<QueryFacilityArgs, 'id'>>;
    provider?: Resolver<Maybe<ResolversTypes['Provider']>, ParentType, ContextType, RequireFields<QueryProviderArgs, 'id'>>;
    providerByNpi?: Resolver<Maybe<ResolversTypes['Provider']>, ParentType, ContextType, RequireFields<QueryProviderByNpiArgs, 'npi'>>;
    providers?: Resolver<Array<ResolversTypes['Provider']>, ParentType, ContextType, Partial<QueryProvidersArgs>>;
    visit?: Resolver<Maybe<ResolversTypes['Visit']>, ParentType, ContextType, RequireFields<QueryVisitArgs, 'id'>>;
    visitsByDateRange?: Resolver<Array<ResolversTypes['Visit']>, ParentType, ContextType, RequireFields<QueryVisitsByDateRangeArgs, 'endDate' | 'startDate'>>;
    visitsForCase?: Resolver<Array<ResolversTypes['Visit']>, ParentType, ContextType, RequireFields<QueryVisitsForCaseArgs, 'caseId'>>;
    visitsForHospital?: Resolver<Array<ResolversTypes['Visit']>, ParentType, ContextType, RequireFields<QueryVisitsForHospitalArgs, 'hospitalId'>>;
    visitsForPatient?: Resolver<Array<ResolversTypes['Visit']>, ParentType, ContextType, RequireFields<QueryVisitsForPatientArgs, 'patientId'>>;
    visitsForProvider?: Resolver<Array<ResolversTypes['Visit']>, ParentType, ContextType, RequireFields<QueryVisitsForProviderArgs, 'providerId'>>;
}>;
export type RecommendationResolvers<ContextType = DataSourceContext, ParentType extends ResolversParentTypes['Recommendation'] = ResolversParentTypes['Recommendation']> = ResolversObject<{
    __resolveReference?: ReferenceResolver<Maybe<ResolversTypes['Recommendation']>, ParentType, ContextType>;
    id?: Resolver<ResolversTypes['ID'], ParentType, ContextType>;
    __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
}>;
// ...
export type Resolvers<ContextType = DataSourceContext> = ResolversObject<{
    Address?: AddressResolvers<ContextType>;
    Case?: CaseResolvers<ContextType>;
    DateTime?: GraphQLScalarType;
    Facility?: FacilityResolvers<ContextType>;
    Hospital?: HospitalResolvers<ContextType>;
    Mutation?: MutationResolvers<ContextType>;
    Patient?: PatientResolvers<ContextType>;
    Provider?: ProviderResolvers<ContextType>;
    Query?: QueryResolvers<ContextType>;
    Recommendation?: RecommendationResolvers<ContextType>;
    Visit?: VisitResolvers<ContextType>;
}>;